+++ router/test.js
@@ -3,7 +3,7 @@ import asyncify from 'express-asyncify'
 import { pgQuery } from '@database/postgres'
 import { configDotenv } from 'dotenv'
 import { useMongoModel } from '../database/mongodb'
-import { testSchema } from '../models/test'
+import { testSchema } from '@model/test'
 import { redisClient } from '@database/redis'
 import { getFromRedis, setToRedis } from '../database/redis'
 import * as validation from '@utility/validation'
